Moved cosmetic to /info command
